+++ trunk/src/team/teams_list.cpp
@@ -196,7 +196,7 @@ void TeamsList::RandomizeFirstPlayer()
 {
   active_team = playing_list.begin();
   MSG_DEBUG("random.get", "TeamList::RandomizeFirstPlayer()");
-  int skip = RandomSync().GetLong(0, playing_list.size() - 1);
+  int skip = RandomSync().GetInt(0, playing_list.size() - 1);
   for (int i = 0; i < skip; i++)
     active_team++;
 }
